AIR-007.2 Refactor React dashboard architecture
