+++ services/dashboard/frontend/src/hooks/useDashboardData.js
@@ -1,0 +1,44 @@
+import { useEffect, useState } from "react";
+
+export function useDashboardData() {
+  const [payload, setPayload] = useState(null);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState("");
+
+  useEffect(() => {
+    let cancelled = false;
+
+    async function loadData() {
+      try {
+        setLoading(true);
+        setError("");
+
+        const response = await fetch("/api/dashboard");
+        if (!response.ok) {
+          throw new Error(`Dashboard API returned ${response.status}`);
+        }
+
+        const data = await response.json();
+        if (!cancelled) {
+          setPayload(data);
+        }
+      } catch (loadError) {
+        if (!cancelled) {
+          setError(loadError instanceof Error ? loadError.message : "Unknown dashboard error");
+        }
+      } finally {
+        if (!cancelled) {
+          setLoading(false);
+        }
+      }
+    }
+
+    loadData();
+
+    return () => {
+      cancelled = true;
+    };
+  }, []);
+
+  return { payload, loading, error };
+}
